Add Company, Person, and ErrorResponse structs
We want to be able to programatically extract more detailed information
from Clearbit API responses, which requires we have data types that map
the API responses.

This adds `Company`, `Person`, and `ErrorResponse` types, and additional
types needed to represent all of the data returned from Clearbit for
these types.

The structs were generated from the fixtures with [gojson], and then
edited for human consumption.

  [gojson]: github.com/ChimeraCoder/gojson/gojson

The only special type introduced here is `TwitterID`. It does some extra
work when unmarshaling from responses to handle the fact that Clearbit
is inconsistent about returning Twitter IDs as integers or strings.