Add EnrichCompany method on Client
This adds an `EnrichCompany` method to the Clearbit client for getting
structured information about a company from its domain name.

This implements the proposal in #9, except that it only accepts a domain
name argument, because we don't currently have any cases where we use
non-streaming calls or pass extra options.

The `enrich` subcommand of the clearbit CLI is also updated to use the
new client API.